requirements.txt
mistralai
python-dotenv
colorama